Extract tests from lib_extensions module
